+++ app/templates/demo/src/app/_app.module.ts
@@ -1,8 +1,9 @@
 
 import { NgModule } from '@angular/core';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { FormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
+import { TransferHttpCacheModule } from '@nguniversal/common';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppSharedModule } from './shared/shared.module';
@@ -17,9 +18,10 @@ import { AppComponent } from './app.component';
         // Add .withServerTransition() to support Universal rendering.
         // The application ID can be any identifier which is unique on
         // the page.
-        BrowserModule.withServerTransition({appId: '<%= projectName %>-demo-id'}),
+        BrowserModule.withServerTransition({ appId: '<%= projectName %>-demo-id' }),
+        TransferHttpCacheModule,
         FormsModule,
-        HttpModule,
+        HttpClientModule,
         AppRoutingModule,
         AppSharedModule,
         HomeModule